requirements.txt
paho-mqtt
pika
requests
prophet
pandas
matplotlib